+++ src/serial_595_lcd.cpp
@@ -4,195 +4,218 @@
 #include <Arduino.h>
 
 /**
-* Performs the signalling required to write to a LCD port
-* Write the raw data to the LCD port shift register.
+* @deprecated
 */
 void lcd_port_write(serial_lcd *lcd) {
-  shiftOut(lcd->pin_ser, lcd->pin_srclk, LSBFIRST, lcd->data);
-  digitalWrite(lcd->pin_rclk, HIGH);
-//  delayMicroseconds(5);
-  digitalWrite(lcd->pin_rclk, LOW);
-  digitalWrite(lcd->pin_ser, LOW);
+  lcd->port_write();
 }
 
+// @deprecated: should use class member function
 void lcd_on(serial_lcd *lcd) {
-  set_bit(&lcd->data, LCD_POWER);
-  lcd_port_write(lcd);
-  delay(20); // we do require a bit more time after power up before we start to initialize this.
-  lcd_initialize(lcd);
-  lcd_backlight_on(lcd);
+  lcd->lcd_on();
 }
 
+// @deprecated: should use class member function
 void lcd_off(serial_lcd *lcd) {
-  //backlight_off();
-  //clear_bit(&port_b, LCD_POWER);
-  // really we need to just put every pin low on the LCD port
-  lcd->data = 0;
-  lcd_port_write(lcd);
+  lcd->lcd_off();
 }
 
+// @deprecated
 void lcd_backlight_on(serial_lcd *lcd) {
-  set_bit(&lcd->data, LCD_BACKLIGHT);
-  lcd_port_write(lcd);
+  lcd->backlight_on();
 }
 
+// @deprecated
 void lcd_backlight_off(serial_lcd *lcd) {
-  clear_bit(&lcd->data, LCD_BACKLIGHT);
-  lcd_port_write(lcd);
+  lcd->backlight_off();
 }
 
+// @deprecated
 void lcd_toggle_backlight(serial_lcd *lcd) {
-  if (lcd->data & LCD_BACKLIGHT) {
-      lcd_backlight_off(lcd);
-  }
-  else {
-      lcd_backlight_on(lcd);
-  }
+  lcd->toggle_backlight();
 }
 
-void lcd_port_toggle_e(serial_lcd *lcd) {
-  // clear_bit(&lcd->data, LCD_E);
-  // lcd_port_write(lcd);
-  // delayMicroseconds(5);
-  set_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
+// @deprecated
+void lcd_initialize(serial_lcd *lcd) {
+  lcd->initialize();
+}
+
+// deprecated
+void lcd_puts(serial_lcd *lcd, char *str) {
+  lcd->puts(str);
+}
+
+// @deprecated
+void lcd_clear(serial_lcd *lcd) {
+  lcd->clear();
+}
+
+/**
+ * @deprecated
+ */
+void lcd_set_dram_address(serial_lcd *lcd, int address) {
+   lcd->set_dram_address(address);
+}
+
+
+/**
+ * @deprecated
+ */
+void lcd_gotoxy(serial_lcd *lcd, int x, int y) {
+  lcd->gotoxy(x,y);
+}
+
+/**
+ * @deprecated
+ */
+void lcd_home(serial_lcd *lcd) {
+  lcd->home();
+}
+
+
+// serial_lcd member functions
+
+void serial_lcd::port_write() {
+  shiftOut(this->pin_ser, this->pin_srclk, LSBFIRST, this->data);
+  digitalWrite(this->pin_rclk, HIGH);
+  digitalWrite(this->pin_rclk, LOW);
+  digitalWrite(this->pin_ser, LOW);
+}
+
+void serial_lcd::port_toggle_e() {
+  set_bit(&this->data, LCD_E);
+  this->port_write();
   delayMicroseconds(5);
-  clear_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
+  clear_bit(&this->data, LCD_E);
+  this->port_write();
 }
 
-// writes 4 bits
-void lcd_write_nibble(serial_lcd *lcd, int RS, int data) {
+void serial_lcd::write_nibble(int RS, int data) {
   // see the write timing diagram
 
   // set up RS
   if (RS) {
-    set_bit(&lcd->data, LCD_RS);
+    set_bit(&this->data, LCD_RS);
   } else {
-    clear_bit(&lcd->data, LCD_RS);
+    clear_bit(&this->data, LCD_RS);
   }
   
   // set the data. Make sure just the lower 8 bits get changed.
-  clear_bit(&lcd->data, LCD_DATA);
-  lcd->data |= (data & LCD_DATA);
-  lcd_port_write(lcd);
+  clear_bit(&this->data, LCD_DATA);
+  this->data |= (data & LCD_DATA);
+  this->port_write();
 
   // set E to high
-  set_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
+  set_bit(&this->data, LCD_E);
+  this->port_write();
   
   // toggle E to LOW. This causes the data to be written to the LCD.
-  clear_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
+  clear_bit(&this->data, LCD_E);
+  this->port_write();
 }
 
-void lcd_write(serial_lcd *lcd, int RS, int data) {
+void serial_lcd::write_byte(int RS, int data) {
   // set up RS
   if (RS) {
-    set_bit(&lcd->data, LCD_RS);
+    set_bit(&this->data, LCD_RS);
   } else {
-    clear_bit(&lcd->data, LCD_RS);
+    clear_bit(&this->data, LCD_RS);
   }
 
   // set E to high
-  set_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
+  set_bit(&this->data, LCD_E);
+  this->port_write();
   
   // set the data. Make sure just the lower 8 bits get changed.
-  clear_bit(&lcd->data, LCD_DATA);
-  lcd->data |= (((data & 0xF0) >> 4) & LCD_DATA);
-  lcd_port_write(lcd);
+  clear_bit(&this->data, LCD_DATA);
+  this->data |= (((data & 0xF0) >> 4) & LCD_DATA);
+  this->port_write();
 
   // toggle E to LOW. This causes the data to be written to the LCD.
-  clear_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
+  clear_bit(&this->data, LCD_E);
+  this->port_write();
 
   // set the data. Make sure just the lower 8 bits get changed.
-  clear_bit(&lcd->data, LCD_DATA);
-  lcd->data |= ((data & 0xF) & LCD_DATA);
+  clear_bit(&this->data, LCD_DATA);
+  this->data |= ((data & 0xF) & LCD_DATA);
 
-  set_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
-
-  // set E to high
-  // set_bit(&lcd->data, LCD_E);
-  // lcd_port_write(lcd);
+  set_bit(&this->data, LCD_E);
+  this->port_write();
   
   // toggle E to LOW. This causes the data to be written to the LCD.
-  clear_bit(&lcd->data, LCD_E);
-  lcd_port_write(lcd);
+  clear_bit(&this->data, LCD_E);
+  this->port_write();
   
-  delayMicroseconds(100);
+  delayMicroseconds(37); // TODO: this was 80. Make sure things still work.
 }
 
+void serial_lcd::lcd_on() {
+  set_bit(&this->data, LCD_POWER);
+  this->port_write();
+  delay(20); // we do require a bit more time after power up before we start to initialize this.
+  lcd_initialize(this);
+  lcd_backlight_on(this);
+}
 
-void lcd_initialize(serial_lcd *lcd) {
-  digitalWrite(lcd->pin_rclk, LOW); // pre-initialize this to low.
+void serial_lcd::lcd_off() {
+  this->data = 0;
+  this->port_write();
+}
+
+void serial_lcd::backlight_on() {
+  set_bit(&this->data, LCD_BACKLIGHT);
+  this->port_write();
+}
+
+void serial_lcd::backlight_off() {
+  clear_bit(&this->data, LCD_BACKLIGHT);
+  this->port_write();
+}
+
+void serial_lcd::toggle_backlight() {
+  if (this->data & LCD_BACKLIGHT) {
+      this->backlight_off();
+  }
+  else {
+      this->backlight_on();
+  }
+}
+
+void serial_lcd::initialize() {
+  digitalWrite(this->pin_rclk, LOW); // pre-initialize this to low.
 
   // https://electronics.stackexchange.com/questions/102245/hd44780-initialization-for-4-bit-mode
   // http://web.alfredstate.edu/faculty/weimandn/lcd/lcd_initialization/lcd_initialization_index.html
   // http://www.farnell.com/datasheets/50586.pdf
   
-  lcd_write_nibble(lcd, 0, 0x3);  // function set [ 0 0 1 DL N 0 * * ] ; DL=1 for 8 bit mode, 0 for 4 bit mode ; N=1 for 16:1 mux, 8 for 8:1
+  this->write_nibble(0, 0x3);  // function set [ 0 0 1 DL N 0 * * ] ; DL=1 for 8 bit mode, 0 for 4 bit mode ; N=1 for 16:1 mux, 8 for 8:1
   delayMicroseconds(4500);
-  lcd_port_toggle_e(lcd); // toggle E 2 more times (pass the function set value again twice more basically)
+  this->port_toggle_e(); // toggle E 2 more times (pass the function set value again twice more basically)
   delayMicroseconds(100);
-  lcd_port_toggle_e(lcd);
+  this->port_toggle_e();
   delayMicroseconds(100);
-  lcd_write_nibble(lcd, 0, 0x2); // Function set (Set interface to 4 bits)
+  this->write_nibble(0, 0x2); // Function set (Set interface to 4 bits)
   delayMicroseconds(100);
   
-  lcd_write(lcd, 0, 0x2C); // function set (2 lines, 5x11 font)
-  lcd_write(lcd, 0, 0x08); // display off, cursor off, blink off
-  lcd_write(lcd, 0, 0x01); // clear screen & return cursor home
-  lcd_write(lcd, 0, 0x06); // inc cursor to the right when writing and don't shift the screen
-  lcd_write(lcd, 0, 0x0C); // turn display on
+  this->write_byte(0, 0x2C); // function set (2 lines, 5x11 font)
+  this->write_byte(0, 0x08); // display off, cursor off, blink off
+  this->write_byte(0, 0x01); // clear screen & return cursor home
+  this->write_byte(0, 0x06); // inc cursor to the right when writing and don't shift the screen
+  this->write_byte(0, 0x0C); // turn display on
 
-  clear_bit (&lcd->data, LCD_DATA); // turn off data pins
-  lcd_port_write(lcd);
+  clear_bit (&this->data, LCD_DATA); // turn off data pins
+  this->port_write();
 }
 
-void lcd_puts(serial_lcd *lcd, char *str) {
-  unsigned int i = 0;
-  for (; str[i] != 0; i++) {
-    if (i > 15) {
-      lcd_write(lcd, 0, 0x18); // shift the display
-    }
-    lcd_write(lcd, 1, str[i]);
-  }
-  clear_bit (&lcd->data, LCD_DATA); // turn off data pins
-  clear_bit (&lcd->data, LCD_RS);
-  lcd_port_write(lcd);
+void serial_lcd::home() {
+  this->write_byte(0, 0x02); // go to home position
 }
 
-void lcd_clear(serial_lcd *lcd) {
-  lcd_write(lcd, 0, 0x01);
-  delay(1);
+void serial_lcd::set_dram_address(int address) {
+   this->write_byte(0, 0x80 + (address & 0x7F));
 }
 
-// Set DDRAM address in address counter
-// this function takes 37 us
-void lcd_set_dram_address(serial_lcd *lcd, int address) {
-   lcd_write(lcd, 0, 0x80 + (address & 0x7F));
-}
-
-
-// The devices provide a 7 bit DRAM address
-// Internally there are 40 character columns, even if the display does not physically have this many columns
-// The row offsets are not consecutive starting points.
-// e.g. for 40 characters, this only uses 0x00-0x27 (0-39)for the first line
-// but the second line in the dram address starts at 0x40-0x67
-// what about those 25 (0x19) bytes from 0x28-0x39?
-
-// to make it more complicated, 4 line display modules that have less than 40 characters per line
-// e.g. like these 20x4 modles I have http://www.newhavendisplay.com/specs/NHD-0420DZ-NSW-BBW.pdf
-// use the not visible DRAM address positions from line 1 are mapped to display line 3.
-// I guess this means we can not do the scroll display
-
-// The set dram address command is 0x80, plus the memory offset.
-// so we want to convert the input of (x,y) to the linear value
-// that satisfies this mess.
-void lcd_gotoxy(serial_lcd *lcd, int x, int y) {
+void serial_lcd::gotoxy(int x, int y) {
   int offset = 0;
   if (x > 19) {
     // error
@@ -209,12 +232,33 @@ void lcd_gotoxy(serial_lcd *lcd, int x, int y) {
     offset += 20;
   }
   offset += x;
-  lcd_set_dram_address(lcd, offset);
+  this->set_dram_address(offset);
 }
 
-// Set DDRAM address to "00H" from AC and return cursor to its original position if shifted.
-// The contents of DDRAM are not changed.
-// This function takes 1.52ms
-void lcd_home(serial_lcd *lcd) {
-  lcd_write(lcd, 0, 0x02); // go to home position
+void serial_lcd::puts(char *str) {
+  unsigned int i = 0;
+  for (; str[i] != 0; i++) {
+    // TODO: define auto shift ability?
+    // if (i > 15) {
+    //   this->write_byte(0, 0x18); // shift the display
+    // }
+    this->write_byte(1, str[i]);
+  }
+  clear_bit (&this->data, LCD_DATA); // turn off data pins
+  clear_bit (&this->data, LCD_RS);
+  this->port_write();
+}
+
+void serial_lcd::clear() {
+  this->write_byte(0, 0x01);
+  delay(1); // TODO: determine if we can use smaller value here.
+}
+
+void serial_lcd::display_screen(lcd_screen_buffer *screen) {
+    // TODO
+}
+
+// screen_buffer member functions
+void lcd_screen_buffer::clear_screen() {
+  printf("%*s", LCD_DRAM_MAX_COLUMN * LCD_DRAM_MAX_ROW, ' ');
 }